requirements.txt
alembic==1.12.1
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
attrs==24.2.0
beautifulsoup4==4.12.2
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.95.2
google-auth==2.38.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.5
Mako==1.2.4
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==24.0
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.21
pydantic_core==2.14.6
PySocks==1.7.1
python-dotenv==0.21.1
requests==2.31.0
rsa==4.9
selenium==4.9.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.38
starlette==0.27.0
tqdm==4.67.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
urllib3==1.26.20
uvicorn==0.22.0
webdriver-manager==4.0.2
wsproto==1.2.0
zipp==3.15.0
